backend/requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
scikit-learn
python-dotenv
google-generativeai
supabase
